% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_bootstrapping.r
\name{boot_ESS}
\alias{boot_ESS}
\title{Estimate bootstrapping ESS}
\usage{
boot_ESS(
  new_data,
  WESS,
  n_rep,
  b_prop,
  replace = T,
  keep.sample = F,
  empirical = F,
  n_level
)
}
\description{
Estimate bootstrapping ESS
}
