% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_boot_excel.r
\name{boot_excel}
\alias{boot_excel}
\title{Get an excel file for bootstrapping results ready for downloading}
\usage{
boot_excel(bootreport)
}
\description{
Get an excel file for bootstrapping results ready for downloading
}
