% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_bootstrapping.r
\name{boot_plot}
\alias{boot_plot}
\title{Plots for bootstrapping results}
\usage{
boot_plot(boot_res, est_data, selected_cp, information, font_size = 14)
}
\description{
Plots for bootstrapping results
}
