% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cutscore.r
\name{calCountWeight}
\alias{calCountWeight}
\title{Calculate counts and weights across all levels}
\usage{
calCountWeight(new_data, n_cut = 3, d_alpha = 1, SD = 1, empirical = T, EC = 0)
}
\description{
Calculate counts and weights across all levels
}
