% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_splines.r
\name{calGamMinimun}
\alias{calGamMinimun}
\title{Calculate the minimun value of GAM function}
\usage{
calGamMinimun(ess_fit, gam_fit, loc_name)
}
\arguments{
\item{ess_fit}{A data frame of fitted ESS-Count or ESS-Weight results.}

\item{gam_fit}{A fitted GAM object, usually from `mgcv::gam()`.}

\item{loc_name}{A character string giving the column name of the cut-score location variable.}
}
\value{
A named numeric vector with the estimated minimum location `x` and value `y`.
}
\description{
Calculate the minimun value of GAM function
}
