% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_splines.r
\name{calGamOutcome}
\alias{calGamOutcome}
\title{Calculate the outcome of GAM function given x}
\usage{
calGamOutcome(data, par = 1)
}
\arguments{
\item{data}{a gam class}

\item{par}{a numeric value for x-axis value}
}
\description{
Calculate the outcome of GAM function given x
}
