% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_data_ready.r
\name{data_ready}
\alias{data_ready}
\title{data_ready}
\usage{
data_ready(imprt_data)
}
\arguments{
\item{imprt_data}{a list of imported data using \code{\link{read_data}}}
}
\value{
a list of the data ready for the analysis.
}
\description{
data_ready
}
\examples{
\dontrun{
require(embededss)
filePath <- "data/freq_data.xlsx"
imported_data <- read_data(filePath)
data_list <- data_ready(imported_data)
}
}
