% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cutscore.r
\name{.calCW}
\alias{.calCW}
\title{Calculate counts and weights for a sinlge level}
\usage{
.calCW(new_data, d_alpha = 1, cut_level = 2, SD = 1, empirical = T, EC = 0)
}
\description{
Calculate counts and weights for a sinlge level
}
