% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab1_table.r
\name{dt_table_out_med}
\alias{dt_table_out_med}
\title{Generate DataTable for median ESS results}
\usage{
dt_table_out_med(tab1_indi_res, table_options)
}
\arguments{
\item{tab1_indi_res}{a data frame}
}
\description{
Generate DataTable for median ESS results
}
