% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmStanS-package.r
\docType{package}
\name{EmStanS-package}
\alias{EmStanS-package}
\title{Embedded Standard Setting}
\description{
Embedded Standard Setting
}
\details{
Analysis of Embedded standard setting (EmStanS)
}
\author{
Sooyong Lee \email{sooyongl09@utexas.edu}
}
\keyword{package}
