% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_update_tab.r
\name{estCutScore_manual}
\alias{estCutScore_manual}
\title{Estimate ESS cut scores based on user-given cut scores}
\usage{
estCutScore_manual(inp_data, manual_cutpoint, information)
}
\description{
Estimate ESS cut scores based on user-given cut scores
}
