% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab3_function.r
\name{gen_page_data}
\alias{gen_page_data}
\title{gen_page_data}
\usage{
gen_page_data(
  ep,
  datainp,
  fourthData,
  fifthData,
  pageName,
  locName,
  targetLoc,
  lvnm
)
}
\description{
gen_page_data
}
