% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab0_function.r
\name{gen_tab0}
\alias{gen_tab0}
\title{ESS analysis for individual panelists}
\usage{
gen_tab0(information)
}
\arguments{
\item{information}{a list of data generated by \code{\link{get_data_info}}}
}
\value{
a list of data for each panelist about ESS results.
}
\description{
ESS analysis for individual panelists
}
