% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab1_function.r
\name{gen_tab1}
\alias{gen_tab1}
\title{ESS analysis for overall GCA}
\usage{
gen_tab1(tab0, information)
}
\arguments{
\item{tab0}{a list of data coming from \code{\link{gen_tab0}}}

\item{information}{a list of data coming from \code{\link{get_data_info}}}
}
\value{
a list of data for each GCA aggregating all the panelists and the tables about ESS results.
}
\description{
ESS analysis for overall GCA
}
