% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab2_function.r
\name{gen_tab2}
\alias{gen_tab2}
\title{Detailed ESS results}
\usage{
gen_tab2(tab1, information)
}
\arguments{
\item{tab1}{a list of data coming from \code{\link{gen_tab1}}}

\item{information}{a list of data coming from \code{\link{get_data_info}}}
}
\value{
a list containing detailed ESS results
}
\description{
Detailed ESS results
}
