% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab3_function.r
\name{gen_tab3}
\alias{gen_tab3}
\title{Cut score summary}
\usage{
gen_tab3(tab1, information)
}
\arguments{
\item{tab1}{a list of data coming from \code{\link{gen_tab1}}}

\item{information}{a list of data coming from \code{\link{get_data_info}}}
}
\value{
a list containing cut score summaries for each domain
}
\description{
Cut score summary
}
