% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab4_function.r
\name{gen_tab4}
\alias{gen_tab4}
\title{Item review table}
\usage{
gen_tab4(tab1, tab2, information)
}
\arguments{
\item{tab1}{a list of data coming from \code{\link{gen_tab1}}}

\item{tab2}{a list of data coming from \code{\link{gen_tab2}}}

\item{information}{a list of data coming from \code{\link{get_data_info}}}
}
\value{
a list containing item review tables
}
\description{
Item review table
}
