% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{genFakeData}
\alias{genFakeData}
\title{Generate a single data set containing ALD, OOD, and location given a correlation value}
\usage{
genFakeData(fun, cor_value, nlevel, ...)
}
\description{
Generate a single data set containing ALD, OOD, and location given a correlation value
}
