% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_simulation_function.r
\name{genFakeDataSet}
\alias{genFakeDataSet}
\title{generate fake data}
\usage{
genFakeDataSet(
  ngca,
  ntable,
  npanelist,
  cor_val,
  nlevel,
  sdinp = 1,
  ecinp = 0,
  n,
  ...
)
}
\arguments{
\item{ngca}{a numeric indicating the number of GCA}

\item{ntable}{a numeric indicating the number of tables per GCA}

\item{npanelist}{a numeric indicating the number of panelists per table}

\item{cor_val}{a numeric indicating the correlation between locations and ALD}

\item{nlevel}{a numeric indicating the number of levels}

\item{sdinp}{a numeric indicating standard deviation adjusting ESS weights}

\item{ecinp}{a numeric for essencially consistent items}

\item{n}{a numeric for items}

\item{...}{a vector of arguments for \code{runif}}
}
\description{
generate fake data
}
