% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_data_ready.r
\name{get_data_info}
\alias{get_data_info}
\title{get_data_info}
\usage{
get_data_info(data_list, ...)
}
\arguments{
\item{data_list}{a list of prepared data that is generated by \code{\link{data_ready}}}

\item{...}{input information for ESS analysis including gca, ald, location, WESS, gamest, domain, select_domain. Those come from the server inputs.}
}
\value{
a list containing all the information necessary for ESS estimation.
}
\description{
get_data_info
}
\examples{
\dontrun{
require(embededss)
filePath <- "data/freq_data.xlsx"
imported_data <- read_data(filePath)
data_list <- data_ready(imported_data)
data_information <- get_data_info(data_list, grade = c("M3"), ald = "ALD", location = "Loc_RP60", wess = F, modal = F, threshold = F)

}
}
