% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_operationlevel.r
\name{getCor}
\alias{getCor}
\title{Estimate the operational levels and the correlation}
\usage{
getCor(inp_data, selected_CP, lvname)
}
\description{
Estimate the operational levels and the correlation
}
