% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_runAPP.r
\name{launchEmStanS}
\alias{launchEmStanS}
\title{Launch EmStanS shiny app}
\usage{
launchEmStanS(local = T)
}
\arguments{
\item{local}{a logical indicating whether the app is launched by the user's local
computer or online. (default = \code{FALSE})}
}
\description{
An interactive Shiny application.
}
\details{
\code{\link{launchEmStanS}} is a caller function to open the Shiny interface of embedded standard setting.

This starts the EmStanS Shiny application on the user's local computer
or launch the online application.
}
\examples{
\dontrun{
if (interactive()) {
  launchEmStanS()
}
}
}
\keyword{Embedded}
\keyword{Setting}
\keyword{Standard}
