% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab1_item_plots.r
\name{plot_dist_items}
\alias{plot_dist_items}
\title{Generate item distribution plots}
\usage{
plot_dist_items(information, x_name = "", y_name = "", font_size = 16)
}
\arguments{
\item{information}{a list of data generated by \code{\link{get_data_info}}}

\item{x_name}{a character for y-axis title}

\item{font_size}{an integer for font size of plot}
}
\value{
a list of item distribution plots for each GCA
}
\description{
Generate item distribution plots
}
