% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plotting}
\alias{plotting}
\title{Plot EmstanS plot}
\usage{
plotting(
  dataUse_1,
  selected_CP,
  selected_CS,
  selected_weights,
  WESS,
  GAM,
  gam_est = NA,
  font_size = 16
)
}
\description{
Plot EmstanS plot
}
