% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_data_ready.r
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(filePath, sheets_name)
}
\arguments{
\item{filePath}{a character for a file path.}

\item{sheets_name}{a list for names of excel sheets}
}
\value{
a named list containing standard setting information (excel data with 5 sheets.)
}
\description{
read_data
}
\examples{
\dontrun{
filePath <- "data/freq_data.xlsx"
imported_data <- read_data(filePath)
}
}
