% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.r
\name{report_itemtable2}
\alias{report_itemtable2}
\title{report_itemtable2}
\usage{
report_itemtable2(imprt_data)
}
\arguments{
\item{imprt_data}{a list of data from \code{\link{read_data}}}
}
\value{
kable output
}
\description{
The function provides the table for the number of items in the item meta data when the data is imported.
}
