% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.r
\name{report_setting}
\alias{report_setting}
\title{report_setting}
\usage{
report_setting(imprt_data)
}
\arguments{
\item{imprt_data}{a list of data from \code{\link{read_data}}}
}
\value{
kable output
}
\description{
The function provides the table for setup data when the data is imported.
}
