% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.r
\name{report_setup}
\alias{report_setup}
\title{report_setup}
\usage{
report_setup(output)
}
\arguments{
\item{output}{ESS object from \code{\link{emstans}}}
}
\value{
html output
}
\description{
report_setup
}
