% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab0_function.r
\name{runESS}
\alias{runESS}
\title{Estimate the cut score for one level.}
\usage{
runESS(new_data, location, n_cut, d_alpha = 1, SD = 1, empirical = T, EC = 0)
}
\description{
Estimate the cut score for one level.
}
