% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nonlinear.r
\name{splineFit}
\alias{splineFit}
\title{Estimate a generalized additive model (GAM) for ESS cut scores}
\usage{
splineFit(data_inp, information = NULL)
}
\description{
Estimate a generalized additive model (GAM) for ESS cut scores
}
