% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab3_function.r
\name{summarize_page}
\alias{summarize_page}
\title{summarize_page}
\usage{
summarize_page(
  ep,
  blankPage,
  perIns,
  percAtabove,
  perBelow,
  minData,
  lvnm,
  numItem,
  cut_score_data
)
}
\description{
summarize_page
}
