% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab1_function.r
\name{tab1_group_out}
\alias{tab1_group_out}
\title{tab1_group_out}
\usage{
tab1_group_out(
  inp_data,
  selectedCp,
  selectedCs,
  selectedWeights,
  WESS,
  modal = F
)
}
\description{
tab1_group_out
}
