% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab1_function.r
\name{tab1_group_out_all}
\alias{tab1_group_out_all}
\title{tab1_group_out_all}
\usage{
tab1_group_out_all(
  inp_data,
  selectedCp,
  selectedCs,
  selectedWeights,
  WESS,
  modal = F
)
}
\description{
tab1_group_out_all
}
