% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab2_function.r
\name{tab2_crosstab}
\alias{tab2_crosstab}
\title{dataUse = tbl_data
EC = EC; distance_res = distance_inp}
\usage{
tab2_crosstab(dataUse, EC = 0, distance_res = NULL, lv_nm = NULL)
}
\description{
dataUse = tbl_data
EC = EC; distance_res = distance_inp
}
