% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab2_function.r
\name{tab2_plot}
\alias{tab2_plot}
\title{tab2_plot}
\usage{
tab2_plot(
  data_inp,
  plot_data,
  information,
  vvv,
  selected_cp,
  selected_cs,
  gam_fitted,
  font_size = 14
)
}
\description{
tab2_plot
}
