% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab2_table.r
\name{tab2_table}
\alias{tab2_table}
\title{Generate DataTable for detailed ESS results}
\usage{
tab2_table(dataInp, information)
}
\arguments{
\item{dataInp}{a data frame}
}
\description{
Generate DataTable for detailed ESS results
}
