% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab2_table.r
\name{tab2_table_crosst}
\alias{tab2_table_crosst}
\title{Generate DataTable for crosstab}
\usage{
tab2_table_crosst(crosstabs)
}
\arguments{
\item{crosstabs}{a data frame}
}
\description{
Generate DataTable for crosstab
}
