% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab2_table.r
\name{tab2_table_effpage}
\alias{tab2_table_effpage}
\title{Generate DataTable for summary table}
\usage{
tab2_table_effpage(dataInp)
}
\arguments{
\item{dataInp}{a data frame}
}
\description{
Generate DataTable for summary table
}
