% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_tab4_table.r
\name{tab4_table_review}
\alias{tab4_table_review}
\title{Generate DataTable obejct for item review table}
\usage{
tab4_table_review(tab4_split)
}
\description{
Generate DataTable obejct for item review table
}
