% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_validate_function.r
\name{validateData}
\alias{validateData}
\title{Check input data for validation}
\usage{
validateData(imprt_data)
}
\arguments{
\item{imprt_data}{a list containing imported data such as setup, panelists, rating, item meta, and examinee data, which comes from \code{\link{read_data}}.}
}
\value{
a list containing messages related to validation.
}
\description{
Check input data for validation
}
