% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmStanS-package.r
\docType{package}
\name{EmStanS-package}
\alias{EmStanS-package}
\title{Embedded Standard Setting}
\description{
Embedded Standard Setting
}
\details{
Embedded Standard Setting (ESS) is an approach to establish
performance standards that augments and is embedded within existing
principled assessment design (PAD) practices. Under ESS, achievement level
descriptor (ALD) writing and the alignment of test items to specified
achievement levels are viewed as the fundamental standard-setting activities
and cut scores are estimated by optimizing the relationship between the
item-ALD alignments and the empirically. Under ESS, this package provides
cut score estimation algorithm (referred to as ESS-Count and ESS-Weight).
}
\keyword{package}
