% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{emstans}
\alias{emstans}
\title{Conduct Embeded Standard Setting}
\usage{
emstans(data, lvname = NULL, WESS = T, GAM = T, SD = 1, EC = 0, digits = 3)
}
\description{
Conduct Embeded Standard Setting
}
\examples{
a1 <- emstans(data, lvname = c("Level1", "Level2", "Level3"))
}
