% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_update_tab.r
\name{estCutScore_mode_manual}
\alias{estCutScore_mode_manual}
\title{Estimate modal ESS cut scores based on user-given cut scores.}
\usage{
estCutScore_mode_manual(data, information, manual_cutpoint)
}
\description{
Estimate modal ESS cut scores based on user-given cut scores.
}
